add global setup and teardown functions for mocha
