+++ server/package.json
@@ -6,11 +6,11 @@
   "scripts": {
     "start": "nodemon src/index.js",
     "database": "rethinkdb",
-    "test": "NODE_ENV=test nyc --all mocha src/test/unit/**/*.js",
-    "test-all": "NODE_ENV=test nyc --all mocha src/test/**/*.js",
+    "test": "NODE_ENV=test nyc --all mocha test/unit/**/*.js",
+    "test-all": "NODE_ENV=test nyc --all mocha test/**/*.js",
     "coverage": "nyc report --reporter=text-lcov > coverage.lcov && codecov",
-    "lint": "eslint src",
-    "fix": "eslint --fix src"
+    "lint": "eslint src test",
+    "fix": "eslint --fix src test"
   },
   "repository": {
     "type": "git",
@@ -47,6 +47,7 @@
     "eslint-config-prettier": "^2.7.0",
     "eslint-plugin-prettier": "^2.3.1",
     "mocha": "^4.0.1",
+    "mocha-prepare": "^0.1.0",
     "nodemon": "^1.12.1",
     "nyc": "^11.3.0",
     "prettier": "^1.7.4",
@@ -58,9 +59,7 @@
       "**/src/**/*.js"
     ],
     "exclude": [
-      "**/src/utils/seed.js",
-      "**/src/utils/mockObjects.js",
-      "**/src/test/**"
+      "**/src/utils/seed.js"
     ]
   }
 }
