+++ server/test/integration/badEndpointTest.js
@@ -1,18 +1,14 @@
-const app = require('../../server');
+const app = require('../../src/server');
 const request = require('supertest');
 const expect = require('chai').expect;
-const axios = require('axios');
-const testObjects = require('../../utils/testObjects');
-const strings = require('../../utils/stringConstants');
+const strings = require('../../src/utils/stringConstants');
 
 // this test suite is for routes unhandled by the API routers
 describe(`'bad' routes - api test`, function() {
-  let token;
+  const token = global.testToken;
 
   before(async function() {
     this.timeout(9000);
-    const response = await axios(testObjects.tokenRequestOptions);
-    token = response.data.access_token;
     return;
   });
